Move remote pull logic into its own file
