+++ server/src/sync/translation_remote/shipment_line.rs
@@ -6,11 +6,12 @@ use repository::{
 
 use serde::Deserialize;
 
-use crate::sync::{translation_remote::IntegrationUpsertRecord, SyncTranslationError};
+use crate::sync::SyncTranslationError;
 
 use super::{
-    empty_str_as_option, zero_date_as_option, IntegrationRecord, RemotePullTranslation,
-    TRANSLATION_RECORD_TRANS_LINE,
+    empty_str_as_option,
+    pull::{IntegrationRecord, IntegrationUpsertRecord, RemotePullTranslation},
+    zero_date_as_option, TRANSLATION_RECORD_TRANS_LINE,
 };
 
 #[derive(Deserialize, Debug)]
@@ -60,7 +61,7 @@ impl RemotePullTranslation for ShipmentLineTranslation {
         &self,
         connection: &StorageConnection,
         sync_record: &RemoteSyncBufferRow,
-    ) -> Result<Option<super::IntegrationRecord>, SyncTranslationError> {
+    ) -> Result<Option<IntegrationRecord>, SyncTranslationError> {
         let table_name = TRANSLATION_RECORD_TRANS_LINE;
         if sync_record.table_name != table_name {
             return Ok(None);
