+++ server/src/sync/translation_remote/stocktake_line.rs
@@ -8,8 +8,9 @@ use serde::Deserialize;
 use crate::sync::SyncTranslationError;
 
 use super::{
-    empty_str_as_option, zero_date_as_option, IntegrationRecord, IntegrationUpsertRecord,
-    RemotePullTranslation, TRANSLATION_RECORD_STOCKTAKE_LINE,
+    empty_str_as_option,
+    pull::{IntegrationRecord, IntegrationUpsertRecord, RemotePullTranslation},
+    zero_date_as_option, TRANSLATION_RECORD_STOCKTAKE_LINE,
 };
 
 #[allow(non_snake_case)]
@@ -44,7 +45,7 @@ impl RemotePullTranslation for StocktakeLineTranslation {
         &self,
         _: &StorageConnection,
         sync_record: &RemoteSyncBufferRow,
-    ) -> Result<Option<super::IntegrationRecord>, SyncTranslationError> {
+    ) -> Result<Option<IntegrationRecord>, SyncTranslationError> {
         let table_name = TRANSLATION_RECORD_STOCKTAKE_LINE;
 
         if sync_record.table_name != table_name {
