+++ server/src/sync/translation_remote/test_data/stocktake.rs
@@ -4,7 +4,8 @@ use repository::schema::{
 };
 
 use crate::sync::translation_remote::{
-    test_data::TestSyncRecord, IntegrationRecord, IntegrationUpsertRecord,
+    pull::{IntegrationRecord, IntegrationUpsertRecord},
+    test_data::TestSyncRecord,
     TRANSLATION_RECORD_STOCKTAKE,
 };
